Module arrow_json::reader::schema

Structs

ValueIter

JSON file reader that produces a serde_json::Value iterator from a Read trait

Enums

InferredType 🔒

Functions

coerce_data_type 🔒

Coerce data type during inference

collect_field_types_from_object 🔒

generate_datatype 🔒

generate_fields 🔒

generate_schema 🔒

Generate schema from JSON field names and inferred data types

infer_array_element_type 🔒

infer_json_schema

Infer the fields of a JSON file by reading the first n records of the buffer, with max_read_records controlling the maximum number of records to read.

infer_json_schema_from_iterator

Infer the fields of a JSON file by reading all items from the JSON Value Iterator.

infer_json_schema_from_seekable

Infer the fields of a JSON file by reading the first n records of the file, with max_read_records controlling the maximum number of records to read.

infer_nested_array_type 🔒

infer_scalar_array_type 🔒

infer_struct_array_type 🔒

set_object_scalar_field_type 🔒