Module arrow_json::reader::tape

Macros

next 🔒

Evaluates to the next element in the iterator or breaks the current loop

Structs

BufIter 🔒

A wrapper around a slice iterator that provides some helper functionality

Tape

A decoded JSON tape

TapeDecoder

Implements a state machine for decoding JSON to a tape

Enums

DecoderState 🔒

States based on https://www.json.org/json-en.html

Literal 🔒

TapeElement

We decode JSON to a flattened tape representation, allowing for efficient traversal of the JSON data

Functions

char_from_surrogate_pair 🔒

Creates a character from an UTF-16 surrogate pair

err 🔒

Returns an error for a given byte b and context ctx

json_whitespace 🔒

Evaluates to true if b is a valid JSON whitespace character

parse_hex 🔒

Parse a hex character to u8

write_char 🔒

Writes c as UTF-8 to out