arrow_json::writer

Module encoder

Macros

float_encode 🔒

integer_encode 🔒

Structs

BinaryEncoder 🔒

New-type wrapper for encoding the binary types in arrow: Binary, LargeBinary and FixedSizeBinary as hex strings in JSON.

BinaryViewEncoder 🔒

BooleanEncoder 🔒

DictionaryEncoder 🔒

EncoderOptions

Configuration options for the JSON encoder.

FieldEncoder 🔒

FixedSizeListEncoder 🔒

JsonArrayFormatter 🔒

A newtype wrapper around [ArrayFormatter] to keep our usage of it private and not implement Encoder for the public type

ListEncoder 🔒

MapEncoder 🔒

NullEncoder 🔒

NullableEncoder

An encoder + a null buffer. This is packaged together into a wrapper struct to minimize dynamic dispatch for null checks.

PrimitiveEncoder 🔒

RawArrayFormatter 🔒

A newtype wrapper around JsonArrayFormatter that skips surrounding the value with "

RunEndEncodedEncoder 🔒

StringEncoder 🔒

StringViewEncoder 🔒

StructArrayEncoder 🔒

Traits

Encoder

A trait to format array values as JSON values

EncoderFactory

A trait to create custom encoders for specific data types.

PrimitiveEncode 🔒

Functions

encode_binary 🔒

encode_string 🔒

make_encoder

Creates an encoder for the given array and field.