arrow_json::writer::encoder

Function make_encoder

pub fn make_encoder<'a>(
    field: &'a FieldRef,
    array: &'a dyn Array,
    options: &'a EncoderOptions,
) -> Result<NullableEncoder<'a>, ArrowError>

Creates an encoder for the given array and field.

This first calls the EncoderFactory if one is provided, and then falls back to the default encoders.