Trait arrow_json::writer::encoder::Encoder

pub trait Encoder {
    // Required method
    fn encode(&mut self, idx: usize, out: &mut Vec<u8>);
}

A trait to format array values as JSON values

Nullability is handled by the caller to allow encoding nulls implicitly, i.e. {} instead of {"a": null}

Required Methods

fn encode(&mut self, idx: usize, out: &mut Vec<u8>)

Encode the non-null value at index idx to out

The behaviour is unspecified if idx corresponds to a null index