Crate arrow_json

Transfer data between the Arrow memory format and JSON line-delimited records. See the module level documentation for the reader and writer for usage examples.

Re-exports

pub use self::reader::Reader;

pub use self::reader::ReaderBuilder;

pub use self::writer::ArrayWriter;

pub use self::writer::LineDelimitedWriter;

pub use self::writer::Writer;

Modules

reader

JSON Reader

writer

JSON Writer

Structs

RawDecoder

A low-level interface for reading JSON data from a byte stream

RawReader

Reads JSON data with a known schema directly into arrow RecordBatch

RawReaderBuilder

A builder for RawReader and RawDecoder

Traits

JsonSerializable

Trait declaring any type that is serializable to JSON. This includes all primitive types (bool, i32, etc.).