Trait arrow_json::reader::primitive_array::ParseJsonNumber

trait ParseJsonNumber: Sized {
    // Required method
    fn parse(s: &[u8]) -> Option<Self>;
}

A trait for JSON-specific primitive parsing logic

According to the specification unquoted fields should be parsed as a double-precision floating point numbers, including scientific representation such as 2e3

In practice, it is common to serialize numbers outside the range of an f64 and expect them to round-trip correctly. As such when parsing integers we first parse as the integer and fallback to parsing as a floating point if this fails

Required Methods

fn parse(s: &[u8]) -> Option<Self>

Object Safety

This trait is not object safe.