Struct arrow_json::reader::tape::TapeDecoder

pub struct TapeDecoder {
    elements: Vec<TapeElement>,
    num_rows: usize,
    batch_size: usize,
    bytes: Vec<u8>,
    offsets: Vec<usize>,
    stack: Vec<DecoderState>,
}

Implements a state machine for decoding JSON to a tape

Fields
elements: Vec<TapeElement>num_rows: usizebatch_size: usize

Number of rows to read per batch

bytes: Vec<u8>

A buffer of parsed string data

Note: if part way through a record, i.e. stack is not empty, this may contain truncated UTF-8 data

offsets: Vec<usize>

Offsets into data

stack: Vec<DecoderState>

A stack of DecoderState

Implementations

impl TapeDecoder

pub fn new(batch_size: usize, num_fields: usize) -> Self

Create a new TapeDecoder with the provided batch size and an estimated number of fields in each row

pub fn decode(&mut self, buf: &[u8]) -> Result<usize, ArrowError>

pub fn serialize<S: Serialize>(&mut self, rows: &[S]) -> Result<(), ArrowError>

Writes any type that implements Serialize into this TapeDecoder

pub fn finish(&self) -> Result<Tape<'_>, ArrowError>

Finishes the current Tape

pub fn clear(&mut self)

Clears this TapeDecoder in preparation to read the next batch