arrow_json::writer
Trait JsonFormat

pub trait JsonFormat: Debug + Default {
    // Provided methods
    fn start_stream<W: Write>(&self, _writer: &mut W) -> Result<(), ArrowError> { ... }
    fn start_row<W: Write>(
        &self,
        _writer: &mut W,
        _is_first_row: bool,
    ) -> Result<(), ArrowError> { ... }
    fn end_row<W: Write>(&self, _writer: &mut W) -> Result<(), ArrowError> { ... }
    fn end_stream<W: Write>(&self, _writer: &mut W) -> Result<(), ArrowError> { ... }
}

This trait defines how to format a sequence of JSON objects to a byte stream.

Provided Methods

fn start_stream<W: Write>(&self, _writer: &mut W) -> Result<(), ArrowError>

write any bytes needed at the start of the file to the writer

fn start_row<W: Write>( &self, _writer: &mut W, _is_first_row: bool, ) -> Result<(), ArrowError>

write any bytes needed for the start of each row

fn end_row<W: Write>(&self, _writer: &mut W) -> Result<(), ArrowError>

write any bytes needed for the end of each row

fn end_stream<W: Write>(&self, _writer: &mut W) -> Result<(), ArrowError>

write any bytes needed for the start of each row

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.